Introduce the common package
