+++ sematic/ui/packages/common/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@sematic/common",
+  "version": "0.2.0",
+  "description": "common components",
+  "scripts": {
+    "lint": "eslint --ext .ts ."
+  },
+  "devDependencies": {
+  },
+  "author": "",
+  "license": "ISC"
+}
