+++ sematic/ui/packages/eslint-config/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    "extends": ["react-app", "react-app/jest"]
+}
