+++ sematic/ui/packages/main/.fork-ts-checkerrc
@@ -1,0 +1,5 @@
+{
+    "typescript": {
+        "build": true
+    }
+}
