+++ sematic/ui/packages/main/src/components/Exception.tsx
@@ -1,5 +1,5 @@
 import { Alert, AlertTitle, Link } from "@mui/material";
-import { ExceptionMetadata } from "../Models";
+import { ExceptionMetadata } from "@sematic/common/src/Models";
 
 const DISCORD_MESSAGE = (
   "For assistance with this error message, you can reach out to us on "
