+++ sematic/ui/packages/main/src/components/GitInfo.tsx
@@ -1,8 +1,8 @@
 import PostAddIcon from "@mui/icons-material/PostAdd";
 import { Box, Link, Tooltip, Typography, useTheme } from "@mui/material";
+import { Resolution } from "@sematic/common/src/Models";
 import { RiGitBranchLine, RiGitCommitLine } from "react-icons/ri";
-import { Resolution } from "../Models";
-import { CopyButton } from "./CopyButton";
+import { CopyButton } from "src/components/CopyButton";
 
 /**
  * Turns the following remote formats:
