+++ sematic/ui/packages/main/src/hooks/appHooks.ts
@@ -1,11 +1,11 @@
+import { User } from "@sematic/common/src/Models";
 import { atomWithStorage } from "jotai/utils";
 import { useContext, useMemo } from "react";
 import useAsync from "react-use/lib/useAsync";
 import AppContext from "../appContext";
-import { ExtractContextType } from "../components/utils/typings";
-import { User } from "../Models";
-import { AuthenticatePayload, EnvPayload } from "../Payloads";
-import { useHttpClient } from "./httpHooks";
+import { ExtractContextType } from "src/components/utils/typings";
+import { useHttpClient } from "../hooks/httpHooks";
+import { AuthenticatePayload, EnvPayload } from "src/Payloads";
 
 export const userAtom = atomWithStorage<User | null>('user', null);
