+++ sematic/ui/packages/main/src/hooks/tests/test_appHooks.cy.tsx
@@ -1,5 +1,5 @@
 import { MemoryRouter } from "react-router-dom";
-import {useAuthentication} from "../appHooks"
+import { useAuthentication } from "@sematic/main/src/hooks/appHooks"
 
 describe('useAutentication hook', () => {
     it("produces app context value if auth is not enabled", () => {
