+++ sematic/ui/packages/main/src/hooks/useMatplotLib.ts
@@ -1,3 +1,5 @@
+///<reference path="../vendor/types/mpld3.d.ts" />
+
 import { draw_figure } from "mpld3";
 
 export function useMatplotLib(elementId: string) {
