+++ sematic/ui/packages/main/src/pipelines/Artifacts.tsx
@@ -1,19 +1,19 @@
-import { Artifact } from "../Models";
-import Alert from "@mui/material/Alert";
-import AlertTitle from "@mui/material/AlertTitle";
-import { renderSummary } from "src/types/Types";
 import {
   Box,
   Table,
   TableBody,
   TableCell,
   TableRow,
-  useTheme,
+  useTheme
 } from "@mui/material";
-import { ErrorBoundary } from "react-error-boundary";
-import { CopyButton } from "../components/CopyButton";
+import Alert from "@mui/material/Alert";
+import AlertTitle from "@mui/material/AlertTitle";
+import { Artifact } from "@sematic/common/src/Models";
 import { useEffect, useMemo, useRef } from "react";
-import { usePipelinePanelsContext } from "../hooks/pipelineHooks";
+import { ErrorBoundary } from "react-error-boundary";
+import { CopyButton } from "src/components/CopyButton";
+import { usePipelinePanelsContext } from "src/hooks/pipelineHooks";
+import { renderSummary } from "src/types/Types";
 
 function ArtifactError(props: { error: Error }) {
   return (
