+++ sematic/ui/packages/main/src/pipelines/RunDetailsPanel.tsx
@@ -1,24 +1,24 @@
 import { Box, Typography } from "@mui/material";
 import { styled } from "@mui/system";
+import { Artifact, Edge, Resolution, Run } from "@sematic/common/src/Models";
 import { useCallback, useContext, useMemo } from "react";
-import { UserContext } from "../appContext";
-import { ActionMenu, ActionMenuItem } from "../components/ActionMenu";
-import CalculatorPath from "../components/CalculatorPath";
-import { CopyButton } from "../components/CopyButton";
-import Docstring from "../components/Docstring";
-import RunId from "../components/RunId";
-import RunStateChip from "../components/RunStateChip";
-import { RunTime } from "../components/RunTime";
-import { SnackBarContext } from "../components/SnackBarProvider";
-import Tags from "../components/Tags";
-import { ExtractContextType } from "../components/utils/typings";
-import { useGraphContext } from "../hooks/graphHooks";
-import { usePipelinePanelsContext, usePipelineRunContext } from "../hooks/pipelineHooks";
-import { Artifact, Edge, Resolution, Run } from "../Models";
-import { fetchJSON } from "../utils";
-import PipelinePanelsContext from "./PipelinePanelsContext";
-import PipelineRunViewContext from "./PipelineRunViewContext";
-import RunTabs, { IOArtifacts } from "./RunTabs";
+import { UserContext } from "src/appContext";
+import { ActionMenu, ActionMenuItem } from "src/components/ActionMenu";
+import CalculatorPath from "src/components/CalculatorPath";
+import { CopyButton } from "src/components/CopyButton";
+import Docstring from "src/components/Docstring";
+import RunId from "src/components/RunId";
+import RunStateChip from "src/components/RunStateChip";
+import { RunTime } from "src/components/RunTime";
+import { SnackBarContext } from "src/components/SnackBarProvider";
+import Tags from "src/components/Tags";
+import { ExtractContextType } from "src/components/utils/typings";
+import { useGraphContext } from "src/hooks/graphHooks";
+import { usePipelinePanelsContext, usePipelineRunContext } from "src/hooks/pipelineHooks";
+import PipelinePanelsContext from "src/pipelines/PipelinePanelsContext";
+import PipelineRunViewContext from "src/pipelines/PipelineRunViewContext";
+import RunTabs, { IOArtifacts } from "src/pipelines/RunTabs";
+import { fetchJSON } from "src/utils";
 
 const StyledText = styled('span')`
   font-size: small;
