+++ sematic/ui/packages/main/src/pipelines/ScrollingLogView.tsx
@@ -1,14 +1,14 @@
-import { useEffect, useCallback, useMemo } from "react";
-import { Alert, Box, Button, LinearProgress, useTheme } from "@mui/material";
 import ArrowCircleDownIcon from '@mui/icons-material/ArrowCircleDown';
-import InfiniteScroll from "react-infinite-scroll-component";
-import { useAccumulateLogsUntilEnd, useLogStream } from "../hooks/logHooks";
-import { usePulldownTrigger, useScrollTracker } from "../hooks/scrollingHooks";
-import { ExceptionMetadata } from "../Models";
-import { Exception, ExternalException } from "../components/Exception";
-import usePrevious from "react-use/lib/usePrevious";
-import { useRunPanelContext, useRunPanelLoadingIndicator } from "../hooks/runDetailsHooks";
+import { Alert, Box, Button, LinearProgress, useTheme } from "@mui/material";
 import { styled } from "@mui/system";
+import { ExceptionMetadata } from "@sematic/common/src/Models";
+import { useCallback, useEffect, useMemo } from "react";
+import InfiniteScroll from "react-infinite-scroll-component";
+import usePrevious from "react-use/lib/usePrevious";
+import { Exception, ExternalException } from "src/components/Exception";
+import { useAccumulateLogsUntilEnd, useLogStream } from "src/hooks/logHooks";
+import { useRunPanelContext, useRunPanelLoadingIndicator } from "src/hooks/runDetailsHooks";
+import { usePulldownTrigger, useScrollTracker } from "src/hooks/scrollingHooks";
 
 const DEFAULT_LOG_INFO_MESSAGE = "No more matching lines";
