+++ sematic/ui/packages/main/tsconfig.json
@@ -2,9 +2,13 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "baseUrl": "./",
+        "noEmit": true
     },
     "include": [
         "src",
         "../tests/global.d.ts"
+    ],
+    "references": [
+        { "path": "../common" },
     ]
 }
