+++ sematic/ui/tsconfig.base.json
@@ -18,7 +18,6 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true,
     "jsx": "react-jsx",
     "types": [
       "cypress",
